rename condenser_api2 to bridge_api
