+++ hive/server/bridge_api/__init__.py
@@ -1,0 +1,1 @@
+"""Hive bridge_api methods and support."""
