+++ hive/server/bridge_api/methods.py
@@ -3,7 +3,7 @@
 # pylint: disable=duplicate-code,too-many-arguments,invalid-name
 
 import hive.server.condenser_api.cursor as cursor
-from hive.server.condenser_api2.objects import load_posts, load_posts_reblogs
+from hive.server.bridge_api.objects import load_posts, load_posts_reblogs
 from hive.server.condenser_api.common import (
     return_error_info,
     valid_account,
