--- hive/server/condenser_api2/__init__.py
@@ -1,1 +1,0 @@
-"""Hive condenser_api2 methods and support."""
